Actualizacion v6
Queda pendiente revisar calculo de caloDiarias, y realizar otros calculos

+++ prjProyectoGym/src/prjproyectogym/cls_funciones.java
@@ -34,7 +34,7 @@ public class cls_funciones {
     //----Validacion de numero-----
     public boolean esNumero(String dato){
         try {
-            Integer.parseInt(dato);
+            Float.parseFloat(dato);
             return true;
         } catch (NumberFormatException datofe) {
             return false;
